correccion del tipo de variable int para peso - Animal - master

+++ tpe/tpe/Animal.java
@@ -3,7 +3,7 @@ package tpe;
 public class Animal extends ElementoGanadero{
 	int id;
 	int edadMeses;
-	double peso;
+	int peso;
 	String raza;
 	String sexo;
 	boolean isCapado;
